Split up files further to have a more clear hierarchy
